+++ openquake/commonlib/logictree.py
@@ -48,6 +48,7 @@ from openquake.hazardlib.imt import from_string
 from openquake.hazardlib import geo, valid, nrml, InvalidFile, pmf
 from openquake.hazardlib.sourceconverter import (
     split_coords_2d, split_coords_3d, SourceGroup)
+from openquake.commonlib import lt
 
 #: Minimum value for a seed number
 MIN_SINT_32 = -(2 ** 31)
@@ -544,92 +545,6 @@ def read_source_groups(fname):
     return src_groups
 
 
-def parse_uncertainty_value(node, branchset):
-    """
-    See superclass' method for description and signature specification.
-
-    Doesn't change source model file name, converts other values to either
-    pair of floats or a single float depending on uncertainty type.
-    """
-    if branchset.uncertainty_type in ('sourceModel', 'extendModel'):
-        return node.text.strip()
-    elif branchset.uncertainty_type == 'abGRAbsolute':
-        [a, b] = node.text.strip().split()
-        return float(a), float(b)
-    elif branchset.uncertainty_type == 'incrementalMFDAbsolute':
-        min_mag, bin_width = (node.incrementalMFD["minMag"],
-                              node.incrementalMFD["binWidth"])
-        return min_mag,  bin_width, ~node.incrementalMFD.occurRates
-    elif branchset.uncertainty_type == 'simpleFaultGeometryAbsolute':
-        return _parse_simple_fault_geometry_surface(node.simpleFaultGeometry)
-    elif branchset.uncertainty_type == 'complexFaultGeometryAbsolute':
-        return _parse_complex_fault_geometry_surface(node.complexFaultGeometry)
-    elif branchset.uncertainty_type ==\
-            'characteristicFaultGeometryAbsolute':
-        surfaces = []
-        for geom_node in node.surface:
-            if "simpleFaultGeometry" in geom_node.tag:
-                trace, usd, lsd, dip, spacing =\
-                    _parse_simple_fault_geometry_surface(geom_node)
-                surfaces.append(geo.SimpleFaultSurface.from_fault_data(
-                    trace, usd, lsd, dip, spacing))
-            elif "complexFaultGeometry" in geom_node.tag:
-                edges, spacing =\
-                    _parse_complex_fault_geometry_surface(geom_node)
-                surfaces.append(geo.ComplexFaultSurface.from_fault_data(
-                    edges, spacing))
-            elif "planarSurface" in geom_node.tag:
-                surfaces.append(
-                    _parse_planar_geometry_surface(geom_node))
-            else:
-                pass
-        if len(surfaces) > 1:
-            return geo.MultiSurface(surfaces)
-        else:
-            return surfaces[0]
-    else:
-        return float(node.text.strip())
-
-    
-def _parse_simple_fault_geometry_surface(node):
-    """
-    Parses a simple fault geometry surface
-    """
-    spacing = node["spacing"]
-    usd, lsd, dip = (~node.upperSeismoDepth, ~node.lowerSeismoDepth,
-                     ~node.dip)
-    # Parse the geometry
-    coords = split_coords_2d(~node.LineString.posList)
-    trace = geo.Line([geo.Point(*p) for p in coords])
-    return trace, usd, lsd, dip, spacing
-
-
-def _parse_complex_fault_geometry_surface(node):
-    """
-    Parses a complex fault geometry surface
-    """
-    spacing = node["spacing"]
-    edges = []
-    for edge_node in node.nodes:
-        coords = split_coords_3d(~edge_node.LineString.posList)
-        edges.append(geo.Line([geo.Point(*p) for p in coords]))
-    return edges, spacing
-
-
-def _parse_planar_geometry_surface(node):
-    """
-    Parses a planar geometry surface
-    """
-    nodes = []
-    for key in ["topLeft", "topRight", "bottomRight", "bottomLeft"]:
-        nodes.append(geo.Point(getattr(node, key)["lon"],
-                               getattr(node, key)["lat"],
-                               getattr(node, key)["depth"]))
-    top_left, top_right, bottom_right, bottom_left = tuple(nodes)
-    return geo.PlanarSurface.from_corner_points(
-        top_left, top_right, bottom_right, bottom_left)
-
-
 class SourceModelLogicTree(object):
     """
     Source model logic tree parser.
@@ -771,7 +686,7 @@ class SourceModelLogicTree(object):
             if validate:
                 self.validate_uncertainty_value(
                     value_node, branchnode, branchset)
-            value = parse_uncertainty_value(value_node, branchset)
+            value = lt.parse_uncertainty_value(value_node, branchset)
             branch_id = branchnode.attrib.get('branchID')
             branch = Branch(bs_id, branch_id, weight, value)
             if branch_id in self.branches:
